third_party/WebKit/LayoutTests/fast/dom/Window/open-as-popup-vs-tab.html

<!doctype html>
<html>
<head>
  <meta charset="utf-8">
  <title>window.open: popup vs tab</title>
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    function openToolbarOnWindow() {
      return window.open("", "", "toolbar=1");
    };
    function openToolbarOffWindow() {
      return window.open("", "", "toolbar=0");
    }
    function openDefaultFeaturesWindow() {
      return window.open();
    }
    test(function() {
      var w = openToolbarOnWindow();
      assert_equals(w.toolbar.visible, true);
      w.close();
    }, "window.open with toolbar=1 creates a new tab");
    test(function() {
      var w = openToolbarOffWindow();
      assert_equals(w.toolbar.visible, false);
      w.close();
    }, "window.open with toolbar=0 creates a new popup");
    test(function() {
      var w = openDefaultFeaturesWindow();
      assert_equals(w.toolbar.visible, true);
      w.close();
    }, "window.open defaults to creating a new tab");
  </script>
  <button onclick="openToolbarOnWindow()">Window Features: toolbar=1</button>
  <button onclick="openToolbarOffWindow()">Window Features: toolbar=0</button>
  <button onclick="openDefaultFeaturesWindow()">Window Features: default</button>
</body>
</html>
